fix: split back the cors
